requirements.txt
Django==1.6.2
Pillow==2.3.0
Unidecode==0.04.14
amqp==1.4.3
anyjson==0.3.3
billiard==3.3.0.16
celery==3.1.9
dj-database-url==0.2.2
dj-static==0.0.5
django-facebook==5.3.1
djangorestframework==2.3.12
gunicorn==18.0
honcho==0.5.0
kombu==3.0.12
psycopg2==2.5.2
pystache==0.5.3
pytz==2013.9
raven==4.0.4
redis==2.9.1
requests==2.2.1
static==1.0.2
wsgiref==0.1.2
